requirements.txt
# Core
pandas>=2.0
numpy>=1.24
google-cloud-bigquery>=3.11
db-dtypes>=1.1

# Visualization
plotly>=5.15
streamlit>=1.28

# Network analysis
networkx>=3.1
python-louvain>=0.16

# Machine learning
scikit-learn>=1.3
scipy>=1.11

# NLP
transformers>=4.35
torch>=2.1
bertopic>=0.15
sentence-transformers>=2.2
langdetect>=1.0.9

# Utilities
tqdm>=4.65
